+++ ni/src/ncl/AddIntrinsics.c
@@ -628,7 +628,7 @@ void _NclAddIntrinsics
 	args[1].dim_sizes[0] = 1;
 	args[1].is_dimsizes = 1;
 	args[1].n_dims = 1;
-	args[2].arg_data_type = _NclLookUp("long");
+	args[2].arg_data_type = _NclLookUp("numeric");
 	args[2].is_dimsizes = 0;
 	args[2].n_dims = 1;
 	args[3].arg_data_type = _NclLookUp("string");
